terminal_ui/requirements.txt
# Terminal UI - Dependencies
PyQt6>=6.6.0
pyqtgraph>=0.13.3
pyserial>=3.5
numpy>=1.24.0
scipy>=1.11.0
google-generativeai>=0.3.0
